DDF-04481 Combined the StsRealm and the WssStsRealm
