--- distribution/docs/src/main/resources/content/_managing/_configuring/external-sts.adoc
@@ -1,38 +1,0 @@
-:title: Connecting to External STS
-:type: subConfiguration
-:status: published
-:parent: Configuring REST Services for Users
-:summary: Configuring to use an existing IdP outside of ${branding}.
-:order: 01
-
-// This is a subsection of Configuring Included IdP and has a added title level.
-=== {title}
-
-Configure ${branding} to connect to an external WSS STS.
-
-.Security STS Address Provider
-Configure the STS address provider to use WSS.
-
-. Navigate to the *${admin-console}*.
-. Select the *${ddf-security}* application.
-. Select *Configuration*.
-. Select the *Security STS Address Provider*.
-. Enable the option *Use WSS STS*.
-
-.Security STS WSS
-Configure the location and credentials for the STS.
-
-. Navigate to the *${admin-console}*.
-. Select the *${ddf-security}* application.
-. Select *Configuration*.
-. Select the *Security STS WSS* configuration.
-. Update the `Address`, `Endpoint Name`, and `Service Name` properties.
-
-.Disable Security STS Client Configuration
-Disable the client configuration for the Security STS
-
-. Navigate to the *${admin-console}*.
-. Select the *System* tab.
-. Select the *Features* tab.
-. Uninstall the *Security STS Client* feature.
-
