--- distribution/docs/src/main/resources/content/_reference/_tables/sts.address.provider.adoc
@@ -1,27 +1,0 @@
-:title: Security STS Address Provider
-:id: ddf.security.sts.address.provider
-:type: table
-:status: published
-:application: ${ddf-security}
-:summary: STS Address Provider.
-
-.[[_ddf.security.sts.address.provider]]Security STS Address Provider
-[cols="1,1m,1,3,1,1" options="header"]
-|===
-
-|Name
-|Id
-|Type
-|Description
-|Default Value
-|Required
-
-|Use WSS STS
-|useWss
-|Boolean
-|If you have a WSS STS configured, you may prefer to use it for services that need the STS address, such as SOAP sources.
-|false
-|true
-
-|===
-
