--- distribution/docs/src/main/resources/content/_reference/_tables/sts.wss.configuration.adoc
@@ -1,97 +1,0 @@
-:title: Security STS WSS
-:id: ddf.security.sts.wss.configuration
-:status: published
-:type: table
-:application: ${ddf-security}
-:summary: STS WSS configurations.
-
-.[[_ddf.security.sts.wss.configuration]]Security STS WSS
-[cols="1,1m,1,3,1,1" options="header"]
-|===
-
-|Name
-|Id
-|Type
-|Description
-|Default Value
-|Required
-
-|SAML Assertion Type
-|assertionType
-|String
-|The version of SAML to use. Most services require SAML v2.0. Changing this value from the default could cause services to stop responding.
-|http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV2.0
-|true
-
-|SAML Key Type
-|keyType
-|String
-|The key type to use with SAML. Most services require Bearer. Changing this value from the default could cause services to stop responding.
-|http://docs.oasis-open.org/ws-sx/ws-trust/200512/Bearer
-|true
-
-|SAML Key Size
-|keySize
-|String
-|The key size to use with SAML. The default key size is 256 and this is fine for most applications. Changing this value from the default could cause services to stop responding.
-|256
-|true
-
-|Use Key
-|useKey
-|Boolean
-|Signals whether or not the STS Client should supply a public key to embed as the proof key. Changing this value from the default could cause services to stop responding.
-|true
-|true
-
-|STS WSDL Address
-|address
-|String
-|STS WSDL Address
-|${variable-prefix}org.codice.ddf.system.protocol}${variable-prefix}org.codice.ddf.system.hostname}:${variable-prefix}org.codice.ddf.system.httpsPort}${variable-prefix}org.codice.ddf.system.rootContext}/SecurityTokenService?wsdl
-|true
-
-|STS Endpoint Name
-|endpointName
-|String
-|STS Endpoint Name.
-|{http://docs.oasis-open.org/ws-sx/ws-trust/200512/}STS_Port
-|false
-
-|STS Service Name
-|serviceName
-|String
-|STS Service Name.
-|{http://docs.oasis-open.org/ws-sx/ws-trust/200512/}SecurityTokenService
-|false
-
-|Signature Properties
-|signatureProperties
-|String
-|Path to Signature crypto properties. This path can be part of the classpath, relative to ${home_directory}, or an absolute path on the system.
-|etc/ws-security/server/signature.properties
-|true
-
-|Encryption Properties
-|encryptionProperties
-|String
-|Path to Encryption crypto properties file. This path can be part of the classpath, relative to ${home_directory}, or an absolute path on the system.
-|etc/ws-security/server/encryption.properties
-|true
-
-|STS Properties
-|tokenProperties
-|String
-|Path to STS crypto properties file. This path can be part of the classpath, relative to ${home_directory}, or an absolute path on the system.
-|etc/ws-security/server/signature.properties
-|true
-
-|Claims
-|claims
-|String
-|Comma-delimited list of claims that should be requested by the STS.
-|http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier,http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress,http://schemas.xmlsoap.org/ws/2005/05/identity/claims/surname,http://schemas.xmlsoap.org/ws/2005/05/identity/claims/givenname,http://schemas.xmlsoap.org/ws/2005/05/identity/claims/role
-|true
-
-|===
-
