+++ platform/security/sts/security-sts-realm/src/main/java/ddf/security/realm/sts/StsAddressProviderImpl.java
@@ -23,34 +23,15 @@ import org.slf4j.LoggerFactory;
 public class StsAddressProviderImpl implements StsAddressProvider {
   private static final Logger LOGGER = LoggerFactory.getLogger(StsAddressProviderImpl.class);
 
-  private boolean useWss = false;
-
   private final STSClientConfiguration internalSts;
 
-  private final STSClientConfiguration wssSts;
-
-  public StsAddressProviderImpl(STSClientConfiguration internalSts, STSClientConfiguration wssSts) {
+  public StsAddressProviderImpl(STSClientConfiguration internalSts) {
     this.internalSts = internalSts;
-    this.wssSts = wssSts;
-  }
-
-  public boolean isUseWss() {
-    return useWss;
-  }
-
-  public void setUseWss(boolean useWss) {
-    this.useWss = useWss;
   }
 
   @Override
   public String getWsdlAddress() {
-    String currentStsAddress;
-    if (useWss) {
-      currentStsAddress = wssSts.getAddress();
-    } else {
-      currentStsAddress = internalSts.getAddress();
-    }
-    return currentStsAddress;
+    return internalSts.getAddress();
   }
 
   @Override
