+++ platform/security/sts/security-sts-realm/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -60,49 +60,10 @@
 
     <service ref="stsRealm" interface="org.apache.shiro.realm.Realm"/>
 
-    <bean id="wssStsRealm" class="ddf.security.realm.sts.WssStsRealm">
-        <property name="contextPolicyManager" ref="contextPolicyManager"/>
-        <cm:managed-properties persistent-id="ddf.security.sts.wss.configuration"
-                               update-strategy="container-managed"/>
-        <property name="address" value="${org.codice.ddf.system.protocol}${org.codice.ddf.system.hostname}:${org.codice.ddf.system.port}${org.codice.ddf.system.rootContext}/SecurityTokenService?wsdl"/>
-        <property name="endpointName"
-                  value="{http://docs.oasis-open.org/ws-sx/ws-trust/200512/}STS_Port"/>
-        <property name="serviceName"
-                  value="{http://docs.oasis-open.org/ws-sx/ws-trust/200512/}SecurityTokenService"/>
-        <property name="signatureProperties" value="${ddf.etc}/ws-security/server/signature.properties"/>
-        <property name="encryptionProperties" value="${ddf.etc}/ws-security/server/encryption.properties"/>
-        <property name="tokenProperties" value="${ddf.etc}/ws-security/server/signature.properties"/>
-        <property name="assertionType"
-                  value="http://docs.oasis-open.org/wss/oasis-wss-saml-token-profile-1.1#SAMLV2.0"/>
-        <property name="keyType" value="http://docs.oasis-open.org/ws-sx/ws-trust/200512/Bearer"/>
-        <property name="keySize" value="256"/>
-        <property name="useKey" value="true"/>
-        <property name="claims">
-            <list>
-                <value>http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier</value>
-                <value>http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress</value>
-                <value>http://schemas.xmlsoap.org/ws/2005/05/identity/claims/surname</value>
-                <value>http://schemas.xmlsoap.org/ws/2005/05/identity/claims/givenname</value>
-                <value>http://schemas.xmlsoap.org/ws/2005/05/identity/claims/role</value>
-            </list>
-        </property>
-    </bean>
-
-    <service ref="wssStsRealm"
-             interface="ddf.security.sts.client.configuration.STSClientConfiguration">
-        <service-properties>
-            <entry key="id" value="wss"/>
-        </service-properties>
-    </service>
-
-    <service ref="wssStsRealm" interface="org.apache.shiro.realm.Realm"/>
-
     <bean id="stsAddressProvider" class="ddf.security.realm.sts.StsAddressProviderImpl">
         <cm:managed-properties persistent-id="ddf.security.sts.address.provider"
                                update-strategy="container-managed"/>
-        <property name="useWss" value="false"/>
         <argument ref="stsRealm"/>
-        <argument ref="wssStsRealm"/>
     </bean>
 
     <service ref="stsAddressProvider"
